feat: :recycle: move starboard manager into Bot
